+++ ChangeLog.md
@@ -19,6 +19,7 @@
     - Define 'signupTextFive' acUserLogin option
     - Define 'signinFieldset' and 'signupFieldset' acUserLogin options
     - Review fieldset style
+    - Define 'signupHaveEmailAddress' and 'signupHaveUsername' acUserLogin options
 
 ### 1.4.0